Create credit card clients
The credit card clients dataset is full of irregularities and incorrect values. It need to replace them with the right values.This dataset contains information on default payments, demographic factors, credit data, history of payment, and bill statements of credit card clients in Taiwan from April 2005 to September 2005. There are 25 features in the dataset.

original dataset from the UCI Machine Learning repository:

https://archive.ics.uci.edu/ml/datasets/default+of+credit+card+clients